split d.ts
