+++ js/ui/grid_core/filtering/utils.js
@@ -26,7 +26,7 @@ function normalizeGroupingLoadOptions(group) {
 
 /**
  *
- * @param {import('../columns_controller/module').ColumnLookup} lookup
+ * @param {import('../columns_controller/types').Lookup} lookup
  */
 export function normalizeLookupDataSource(lookup) {
     let lookupDataSourceOptions;
@@ -44,7 +44,7 @@ export function normalizeLookupDataSource(lookup) {
 
 /**
  *
- * @param {import('../columns_controller/module').LookupColumn} column
+ * @param {import('../columns_controller/types').LookupColumn} column
  * @param {import('../ui.grid_core.data_source_adapter').DataSourceAdapter} dataSource
  * @param {*} filter
  */
